+++ src/video/AsyncDemuxer.h
@@ -40,7 +40,7 @@ namespace avg {
            
             void enableStream(int StreamIndex);
             AVPacket * getPacket(int StreamIndex);
-            void seek(long long DestTime);
+            void seek(double DestTime);
             
         private:
             void waitForSeekDone();
