+++ src/video/IDemuxer.h
@@ -63,7 +63,7 @@ namespace avg {
            
             virtual void enableStream(int StreamIndex) = 0;
             virtual AVPacket * getPacket(int StreamIndex) = 0;
-            virtual void seek(long long DestTime) = 0;
+            virtual void seek(double DestTime) = 0;
             virtual void dump() {};
             
     };
